Improve CLI
Added two flags to change the credential file path and the socket file
path to enable tests regarding the CLI
without interfering with a local nicator installation.
This required to change the flag parsing code and how path are handled.
Exit codes have also been improved to catch more errors with the tests.
Nicator no longer needs to be installed on $PATH.